modification of identifiedObject

+++ src/model/Bill.java
@@ -13,8 +13,9 @@ public class Bill extends IdentifiedObject{
     }
 
     @Override
-    public Integer getId() {
-        return (Integer) super.getId();
+    public int getId() {
+
+        return super.getId();
     }
 
     public Date getBillDate() {
