+++ src/model/Client.java
@@ -25,8 +25,8 @@ public class Client extends IdentifiedObject {
     }
 
     @Override
-    public Integer getId() {
-        return (Integer) super.getId();
+    public int getId() {
+        return super.getId();
     }
 
     public String getNomCli() {
