+++ src/model/Reservation.java
@@ -33,8 +33,9 @@ public class Reservation extends IdentifiedObject{
     }
 
     @Override
-    public Integer getId() {
-        return (Integer) super.getId();
+    public int getId() {
+
+        return super.getId();
     }
 
     public ReservationStatement getReservationStatement() {
